ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Pillow==10.4.0
torch==2.4.0
torchvision==0.19.0
open-clip-torch==2.24.0
faiss-cpu==1.8.0
numpy==1.26.4

